@@ -23,11 +23,13 @@ set(EXECUTABLE_OUTPUT_PATH "build")
 find_program(CLANG_EXECUTABLE clang++)
 
 SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g -ggdb -O2 -pedantic")
-#SET(CMAKE_CXX_STANDARD 11)
-SET(CMAKE_CXX_STANDARD 17)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
+#SET(CMAKE_CXX_STANDARD 17)
+SET(CMAKE_CXX_STANDARD 20)
+SET(CMAKE_CXX_STANDARD_REQUIRED ON)
 # Hide symbols in UDF code by default - UDF functions should be explicitly exported.
 SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fvisibility=hidden -fvisibility-inlines-hidden")
+SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -march=cascadelake")
+
 
 
 # Build the UDA/UDFs into a shared library.
@@ -36,6 +38,11 @@ add_library(udfmathutils SHARED udf-mathutils.cc)
 add_library(udabioutils SHARED uda-bioutils.cc)
 add_library(udf-sero SHARED udf-sero.cc)
 
+set_target_properties(udfmathutils
+    PROPERTIES
+        CXX_STANDARD 17
+        CXX_STANDARD_REQUIRED YES
+)
 # This is an example of how to use the test harness to help develop UDF and UDAs.
 target_link_libraries(udfbioutils ImpalaUdf)
 target_link_libraries(udfbioutils crypto)
@@ -45,7 +52,6 @@ add_executable(uda-bioutils-test uda-bioutils-test.cc)
 target_link_libraries(uda-bioutils-test udabioutils)
 
 target_link_libraries(udfmathutils ImpalaUdf)
-
 target_link_libraries(udf-sero ImpalaUdf)
 
 add_executable(udf-sero-test udf-sero-test.cc)
